pulled out repeated code into an 'action helpers' file
