+++ test/client/actions/action_test_helpers.js
@@ -1,0 +1,31 @@
+import configureMockStore from 'redux-mock-store';
+import thunk from 'redux-thunk';
+
+const middlewares = [thunk];
+export const mockStore = configureMockStore(middlewares);
+
+const localStorageMock = (() => {
+  let localstore = {};
+  return {
+    getItem(key) {
+      return localstore[key];
+    },
+    setItem(key, value) {
+      localstore[key] = value.toString();
+    },
+    clear() {
+      localstore = {};
+    }
+  };
+})();
+
+export function exposeLocalStorageMock() {
+  global.localStorage = localStorageMock;
+}
+
+export function checkObjectProps(expectedProps, obj) {
+  return expectedProps.reduce((accum, curr) => {
+    const propertyExists = Object.prototype.hasOwnProperty.call(obj, curr);
+    return accum && propertyExists;
+  }, true);
+}
